making schema validation optional in config-loader
